fix rag benchmark, update docs
