+++ benchmark/test_rag_interpretation.py
@@ -2,18 +2,13 @@ import inspect
 
 import pytest
 
-import pandas as pd
-
 from .conftest import calculate_test_score
 from .benchmark_utils import (
     get_result_file_path,
     write_results_to_file,
-    benchmark_already_executed,
+    skip_if_already_run,
 )
 from biochatter._misc import ensure_iterable
-
-TASK = "rag_interpretation"
-FILE_PATH = get_result_file_path(TASK)
 
 
 def get_test_data(test_data_rag_interpretation: list) -> tuple:
@@ -36,24 +31,6 @@ def get_test_data(test_data_rag_interpretation: list) -> tuple:
         test_data_rag_interpretation["test_case_purpose"],
         test_data_rag_interpretation["index"],
     )
-
-
-def skip_if_already_run(
-    model_name: str,
-    result_files: dict[str, pd.DataFrame],
-    subtask: str,
-) -> None:
-    """Helper function to check if the test case is already executed.
-
-    Args:
-        model_name (str): The model name, e.g. "gpt-3.5-turbo"
-        result_files (dict[str, pd.DataFrame]): The result files
-        subtask (str): The benchmark subtask test case, e.g. "entities_0"
-    """
-    if benchmark_already_executed(TASK, subtask, model_name, result_files):
-        pytest.skip(
-            f"benchmark {TASK}: {subtask} with {model_name} already executed"
-        )
 
 
 def test_explicit_relevance_of_single_fragments(
@@ -103,7 +80,7 @@ def test_explicit_relevance_of_single_fragments(
         model_name,
         subtask,
         f"{mean_score}/{max};{n_iterations}",
-        FILE_PATH,
+        get_result_file_path(task),
     )
 
 
@@ -153,5 +130,5 @@ def test_implicit_relevance_of_multiple_fragments(
         model_name,
         subtask,
         f"{mean_score}/{max};{n_iterations}",
-        FILE_PATH,
+        get_result_file_path(task),
     )
